# docs/_archive/2026-08-16/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "侦听台改造项目 核心接口契约",
    "version": "2.0.0",
    "description": "核心接口定义（现状接口 + platform 编排目标接口）。约定：全部 RESTful JSON；时间统一 ISO8601；错误统一 { code, message, detail }。"
  },
  "servers": [
    { "url": "http://127.0.0.1:8765", "description": "侦听台 listener" },
    { "url": "http://127.0.0.1:8766", "description": "模块日志 module_log" },
    { "url": "http://127.0.0.1:8781", "description": "模拟集中器 sim_concentrator 独立入口" },
    { "url": "http://127.0.0.1:8790", "description": "platform 一体化工作台（规划）" }
  ],
  "paths": {
...
    "/api/run": {
      "post": {
        "tags": ["platform-编排"],
        "summary": "创建验证批次（Run）并执行",
        "operationId": "createRun",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["scenario_id"],
                "properties": {
                  "scenario_id": { "type": "string", "example": "minute_collect" },
                  "firmware": { "$ref": "#/components/schemas/FirmwareInfo" },
                  "parameters": { "type": "object" },
                  "skip_flash": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Run 已创建", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Run" } } } },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/api/run/{run_id}": {
      "get": {
        "tags": ["platform-编排"],
        "summary": "查询批次状态与报告",
        "operationId": "getRun",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Run 状态",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/run/{run_id}/report": {
      "get": {
        "tags": ["platform-编排"],
        "summary": "获取统一验证报告",
        "operationId": "getRunReport",
        "parameters": [
          { "name": "run_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "统一验证报告",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Report" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/compare": {
      "post": {
        "tags": ["platform-编排"],
        "summary": "期望流程 vs 实际事件流 → 流程比对结果",
        "operationId": "compareFlow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["expected_flow", "actual_events"],
                "properties": {
                  "expected_flow": { "type": "array", "items": { "$ref": "#/components/schemas/ExpectedStep" } },
                  "actual_events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "流程比对结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FlowCompareResult" }
              }
            }
          }
        }
      }
    },
    "/api/feedback": {
      "post": {
        "tags": ["platform-编排"],
        "summary": "由失败结论生成归因反馈文本",
        "operationId": "createFeedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["compare_result"],
                "properties": {
                  "compare_result": { "$ref": "#/components/schemas/FlowCompareResult" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "归因反馈",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "issue": { "type": "string" },
                          "evidence": { "type": "string" },
                          "suggestion": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Event": {
        "type": "object",
        "required": ["type", "time"],
        "properties": {
          "type": { "type": "string" },
          "time": { "type": "string" },
          "label": { "type": "string" },
          "message": { "type": "string" },
          "source_line": { "type": "string" }
        }
      },
      "FirmwareInfo": {
        "type": "object",
        "properties": {
          "version": { "type": "string" },
          "commit": { "type": "string" },
          "flash_file_sha256": { "type": "string" }
        }
      },
...
      "ExpectedStep": {
        "type": "object",
        "required": ["step", "event_type"],
        "properties": {
          "step": { "type": "string" },
          "event_type": { "type": "string" },
          "within_ms": { "type": "integer" },
          "negate": { "type": "boolean", "default": false },
          "optional": { "type": "boolean", "default": false }
        }
      },
      "FlowCompareResult": {
        "type": "object",
        "properties": {
          "steps": { "type": "array", "items": { "type": "object" } },
          "missing": { "type": "array", "items": { "type": "string" } },
          "timeouts": { "type": "array", "items": { "type": "string" } },
          "out_of_order": { "type": "array", "items": { "type": "string" } },
          "negated": { "type": "array", "items": { "type": "string" } },
          "verdict": { "type": "string", "enum": ["pass", "fail"] }
        }
      },
      "Run": {
        "type": "object",
        "properties": {
          "run_id": { "type": "string", "example": "run-20260814-0001" },
          "scenario_id": { "type": "string" },
          "firmware": { "$ref": "#/components/schemas/FirmwareInfo" },
          "status": { "type": "string", "enum": ["pending", "running", "passed", "failed", "inconclusive", "aborted"] },
          "verdict": { "type": "string", "enum": ["pass", "fail", "inconclusive"] },
          "created_at": { "type": "string" },
          "started_at": { "type": "string" },
          "finished_at": { "type": "string" }
        }
      },
      "Report": {
        "type": "object",
        "properties": {
          "run_id": { "type": "string" },
          "firmware": { "$ref": "#/components/schemas/FirmwareInfo" },
          "scenario": { "type": "string" },
          "sources": { "type": "object" },
          "assertions": { "type": "array", "items": { "type": "object" } },
          "flow_compare": { "$ref": "#/components/schemas/FlowCompareResult" },
          "verdict": { "type": "string", "enum": ["pass", "fail", "inconclusive"] },
          "artifacts": { "type": "array", "items": { "type": "string" } },
          "ts": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "detail": { "type": "object" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "资源不存在",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ValidationError": {
        "description": "参数校验失败",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}